+++ ext/couchbase_ext/arithmetic.c
@@ -75,9 +75,10 @@ cb_bucket_arithmetic(int sign, int argc, VALUE *argv, VALUE self)
     lcb_error_t err;
     struct cb_params_st params;
 
-    if (bucket->handle == NULL) {
-        rb_raise(cb_eConnectError, "closed connection");
+    if (!cb_bucket_connected_bang(bucket, sign > 0 ? cb_sym_increment : cb_sym_decrement)) {
+        return Qnil;
     }
+
     memset(&params, 0, sizeof(struct cb_params_st));
     rb_scan_args(argc, argv, "0*&", &params.args, &proc);
     if (!bucket->async && proc != Qnil) {