+++ ext/couchbase_ext/get.c
@@ -223,9 +223,10 @@ cb_bucket_get(int argc, VALUE *argv, VALUE self)
     lcb_error_t err = LCB_SUCCESS;
     struct cb_params_st params;
 
-    if (bucket->handle == NULL) {
-        rb_raise(cb_eConnectError, "closed connection");
+    if (!cb_bucket_connected_bang(bucket, cb_sym_get)) {
+        return Qnil;
     }
+
     memset(&params, 0, sizeof(struct cb_params_st));
     rb_scan_args(argc, argv, "0*&", &params.args, &proc);
     if (!bucket->async && proc != Qnil) {